requirements.txt
phonenumbers==7.5.2
pillow==3.3.1
selenium==2.53.6
wheel==0.24.0
xlwt==1.1.2
